windowManager: Activate new workspace before removing the current one
When removing the current workspace, the active workspace is changed
to the preceding one automatically before we change explicitly to the
last workspace. There is no good reason to change workspaces twice in
this case, we can avoid the first one just by changing to the new
workspace before removing any workspaces.

https://bugzilla.gnome.org/show_bug.cgi?id=709064