Remove cogl2-compatibility.c
This was added in 361bd516f3d678d late during the 1.10 cycle to
contain experimental functions that we should never have made public.
The plan was to remove them once we started working on 1.12 but it
looks like we never got around to doing that. Better late than never!

The header for the file was already removed in 7365c3aa77fe2efc4f5a71.

Reviewed-by: Robert Bragg <robert@linux.intel.com>